refactor: Change code in 'changePizzaSizes' function
Changed code in 'changePizzaSizes' function to improve its performance.
Put the collection of elements with class name 'randomPizzaContainer' outside of the loop, so that it would be queried only once in the function.
Deleted 'determineDx' as it proved to be unuseful.
Also, instead of calculating a size difference, a new percentage width was assigned to every element.