refactor: Modify updatePositions to reach 60FPS
The 'updatePositions()' function was modified to reach 60FPS when scrolling. This function makes the pizzas on the background move as we scroll the sidebar.
Some computations were simplified inside this function, while some initializations were put outside of it. There were also some heavy computations inside the loops that didn't need to be there, as they weren't relying on any element being looped.
Based on an article written by Paul Lewis on HTML5 Rocks about faster animations with RequestAnimationFrame, I used 'requestAnimationFrame' in my code to ensure the app would render at 60 frames per second when the sidebar was scrolled. I copied the code from the last example on that page and adapted it to my needs.
I also reduced the number of moving pizzas that would move in the background to a smaller quantity, roughly similar to what we really see on the screen at any given scroll.